% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{get_manifold}
\alias{get_manifold}
\title{Retrieve a manifold by name}
\usage{
get_manifold(manifold)
}
\arguments{
\item{manifold}{Either the name of a supported manifold (currently only
\code{"torus"}), or an object created with \code{\link{new_manifold}},
which is returned unchanged.}
}
\value{
An object of class \code{"manifold"}.
}
\description{
Resolves the \code{manifold} argument of \code{\link{manifold.dcov.test}}
and \code{\link{choose_base_point}}.
}
\examples{
get_manifold("torus")$name

}
